Videos/2026-03-12_Gemini_Embedding_2_True_Multimodal_Revolution_2026/requirements.txt
google-cloud-aiplatform>=1.50.0
numpy>=1.21.0
Pillow>=9.0.0